ALL TESTS PASS
avail/compiler:
-No longer captures (final) instances of MapDescriptor.Entry for inner classes within Map foreach constructs.

avail/descriptor/AbstractAvailObject.java:
-Short slot accessors now operate on unsigned shorts (via ints).

avail/descriptor/MapDescriptor.java:
-Removed dispatched private field accessors.
-Reuses a MapDescriptor.Entry during iteration.
-Equal maps collapse together.

avail/descriptor/HashedSetBinDescriptor.java:
-No longer eagerly computes bin union type.

avail/descriptor/SetDescriptor.java:
-Equal sets collapse together.

avail/serialization/Serializer.java:
-Corrected topological sort.

avail/test/SerializerTest.java:
-Now also tests tuples, sets, and maps.

+++ Avail/src/com/avail/descriptor/AbstractAvailObject.java
@@ -227,29 +227,31 @@ abstract class AbstractAvailObject
 		final short aByte);
 
 	/**
-	 * Extract a (16-bit signed) {@code short} at the given short-index of the
+	 * Extract a (16-bit unsigned) {@code short} at the given short-index of the
 	 * receiver.
 	 *
 	 * @param e The enumeration value that identifies the base field.
 	 * @param shortIndex The index in bytes (must be even).
-	 * @return The {@code short} found at the given short-index.
+	 * @return The unsigned {@code short} (as an {@code int} found at the given
+	 *         short-index.
 	 */
-	abstract short shortSlotAt (
+	abstract int shortSlotAt (
 		final @NotNull IntegerSlotsEnum e,
 		final int shortIndex);
 
 	/**
-	 * Store the (16-bit signed) {@code short} at the given short-index of the
+	 * Store the (16-bit unsigned) {@code short} at the given short-index of the
 	 * receiver.
 	 *
 	 * @param e The enumeration value that identifies the base field.
 	 * @param shortIndex The index in bytes (must be even).
-	 * @param aShort The {@code short} to store at the given short-index.
+	 * @param aShort The {@code short} to store at the given short-index, passed
+	 *               as an {@code int} for safety.
 	 */
 	abstract void shortSlotAtPut (
 		final @NotNull IntegerSlotsEnum e,
 		final int shortIndex,
-		final short aShort);
+		final int aShort);
 
 	/**
 	 * Extract the (signed 32-bit) integer for the given field {@code enum}
